+++ AgentSessions/Views/Preferences/PreferencesView+General.swift
@@ -471,6 +471,7 @@ private extension PreferencesView {
         case .kimi:        return $kimiAgentEnabled
         case .grok:        return $grokAgentEnabled
         case .qwen:        return $qwenAgentEnabled
+        case .fx:          return $fxAgentEnabled
         }
     }